Introduce a ModelResolutionTrait and move shared model-selection functionality into that and out of PromptBuilder
